修改数据类型

+++ src/main/java/com/education/hjj/bz/util/weixinUtil/vo/RedpackRequestPo.java
@@ -30,7 +30,7 @@ public class RedpackRequestPo implements Serializable {
 	// 用户openid
 	private String re_openid;
 	// 付款金额
-	private BigDecimal total_amount;
+	private Integer total_amount;
 	// 红包发放总人数
 	private Integer total_num;
 	// 红包祝福语
@@ -120,11 +120,11 @@ public class RedpackRequestPo implements Serializable {
 		this.re_openid = re_openid;
 	}
 
-	public BigDecimal getTotal_amount() {
+	public Integer getTotal_amount() {
 		return total_amount;
 	}
 
-	public void setTotal_amount(BigDecimal total_amount) {
+	public void setTotal_amount(Integer total_amount) {
 		this.total_amount = total_amount;
 	}
